Day 5: Binary Boarding
Read a list of boarding passes that tell you how to navigate the aircraft:

 * The aircraft is 8 columns wide by 128 rows
 * First seven characters indicates what halves of the rows to navigate: an F in the first character means the Front half (0..63) whereas a B would mean the Back half (64..127)
 * The last three characters are similar, but for the column
 * The Seat ID is the `(row * 8) + column`

Calculate the highest Seat ID, then work out the missing seat in the middle of the plane and enter that Seat ID.

    pytest .
    ./binary_boarding.py input

Visualising the missing seat was the quickest route to success in this example. Exactly what you'd do on an aircraft.