Day 6: Custom Customs
For the first question, the sum of all unique characters per double-blank-space separation. For the second question, the sum of the intersection of blank-space separated characters per double-blank space separation.

Not my proudest work: a couple of hacks are present:

1. Identify when there is only one character to be returned, to avoid empty sets.
2. Newlines appended to the `input` file to ensure that the last entry is read.

Room for plenty of optimisation, but it is a Sunday…